requirements.txt
# Core Dependencies
python-dotenv==1.0.1
pydantic==2.5.3

# CAD Processing
ezdxf==1.2.0
ifcopenshell==0.8.3.post2

# Geometry and Graph Analysis
shapely==2.0.2
networkx==3.2.1
numpy==1.26.4
scipy==1.12.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.2

# LangChain and RAG
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.7
langchain-google-genai==1.0.8
langgraph==0.2.16
langchain-core==0.2.38
google-generativeai==0.3.2

# Vector Database and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
langchain-huggingface==0.1.0

# PDF Processing
pypdf==3.17.4
pymupdf==1.23.8

# Utilities
tqdm==4.66.1
rich==13.7.0
click==8.1.7

# Jupyter and Notebooks
jupyter==1.0.0
notebook==7.0.7
ipywidgets==8.1.1
ipykernel==6.28.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# API and Web
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# JSON Schema Validation
jsonschema==4.20.0